+++ cmd/tuple/changes_test.go
@@ -44,7 +44,7 @@ func TestReadChangesError(t *testing.T) {
 
 	mockFgaClient.EXPECT().ReadChanges(context.Background()).Return(mockBody)
 
-	_, err := readChanges(mockFgaClient, 5, "document", "")
+	_, err := readChanges(mockFgaClient, 5, "document", "", "")
 	if err == nil {
 		t.Error("Expect error but there is none")
 	}
@@ -82,7 +82,7 @@ func TestReadChangesEmpty(t *testing.T) {
 
 	mockFgaClient.EXPECT().ReadChanges(context.Background()).Return(mockBody)
 
-	output, err := readChanges(mockFgaClient, 5, "document", "")
+	output, err := readChanges(mockFgaClient, 5, "document", "", "")
 	if err != nil {
 		t.Error(err)
 	}
@@ -139,19 +139,25 @@ func TestReadChangesSinglePage(t *testing.T) {
 
 	mockBody := mock_client.NewMockSdkClientReadChangesRequestInterface(mockCtrl)
 
+	sTime, err := time.Parse(time.RFC3339, "2022-01-01T00:00:00Z")
+	if err != nil {
+		t.Error(err)
+	}
+
 	body := client.ClientReadChangesRequest{
-		Type: "document",
+		Type:      "document",
+		StartTime: sTime,
 	}
 	mockBody.EXPECT().Body(body).Return(mockRequest)
 
 	mockFgaClient.EXPECT().ReadChanges(context.Background()).Return(mockBody)
 
-	output, err := readChanges(mockFgaClient, 5, "document", "")
+	output, err := readChanges(mockFgaClient, 5, "document", "2022-01-01T00:00:00Z", "")
 	if err != nil {
 		t.Error(err)
 	}
 
-	expectedOutput := `{"changes":[{"operation":"TUPLE_OPERATION_WRITE","timestamp":"2009-11-10T23:00:00Z","tuple_key":{"object":"document:doc1","relation":"reader","user":"user:user1"}}],"continuation_token":""}` //nolint:lll
+	expectedOutput := `{"changes":[{"operation":"TUPLE_OPERATION_WRITE","timestamp":"2009-11-10T23:00:00Z","tuple_key":{"object":"document:doc1","relation":"reader","user":"user:user1"}}],"continuation_token":""}`
 
 	outputTxt, err := json.Marshal(output)
 	if err != nil {
@@ -232,8 +238,14 @@ func TestReadChangesMultiPages(t *testing.T) {
 	mockBody1 := mock_client.NewMockSdkClientReadChangesRequestInterface(mockCtrl)
 	mockBody2 := mock_client.NewMockSdkClientReadChangesRequestInterface(mockCtrl)
 
+	sTime, err := time.Parse(time.RFC3339, "2022-01-01T00:00:00Z")
+	if err != nil {
+		t.Error(err)
+	}
+
 	body := client.ClientReadChangesRequest{
-		Type: "document",
+		Type:      "document",
+		StartTime: sTime,
 	}
 
 	gomock.InOrder(
@@ -246,12 +258,12 @@ func TestReadChangesMultiPages(t *testing.T) {
 		mockFgaClient.EXPECT().ReadChanges(context.Background()).Return(mockBody2),
 	)
 
-	output, err := readChanges(mockFgaClient, 5, "document", "")
+	output, err := readChanges(mockFgaClient, 5, "document", "2022-01-01T00:00:00Z", "")
 	if err != nil {
 		t.Error(err)
 	}
 
-	expectedOutput := `{"changes":[{"operation":"TUPLE_OPERATION_WRITE","timestamp":"2009-11-10T22:00:00Z","tuple_key":{"object":"document:doc1","relation":"reader","user":"user:user1"}},{"operation":"TUPLE_OPERATION_DELETE","timestamp":"2009-11-10T23:00:00Z","tuple_key":{"object":"document:doc1","relation":"reader","user":"user:user1"}}],"continuation_token":"01GXSA8YR785C4FYS3C0RTG7B2"}` //nolint:lll
+	expectedOutput := `{"changes":[{"operation":"TUPLE_OPERATION_WRITE","timestamp":"2009-11-10T22:00:00Z","tuple_key":{"object":"document:doc1","relation":"reader","user":"user:user1"}},{"operation":"TUPLE_OPERATION_DELETE","timestamp":"2009-11-10T23:00:00Z","tuple_key":{"object":"document:doc1","relation":"reader","user":"user:user1"}}],"continuation_token":"01GXSA8YR785C4FYS3C0RTG7B2"}`
 
 	outputTxt, err := json.Marshal(output)
 	if err != nil {
@@ -303,19 +315,25 @@ func TestReadChangesMultiPagesLimit(t *testing.T) {
 
 	mockBody := mock_client.NewMockSdkClientReadChangesRequestInterface(mockCtrl)
 
+	sTime, err := time.Parse(time.RFC3339, "2022-01-01T00:00:00Z")
+	if err != nil {
+		t.Error(err)
+	}
+
 	body := client.ClientReadChangesRequest{
-		Type: "document",
+		Type:      "document",
+		StartTime: sTime,
 	}
 	mockBody.EXPECT().Body(body).Return(mockRequest)
 
 	mockFgaClient.EXPECT().ReadChanges(context.Background()).Return(mockBody)
 
-	output, err := readChanges(mockFgaClient, 1, "document", "")
+	output, err := readChanges(mockFgaClient, 1, "document", "2022-01-01T00:00:00Z", "")
 	if err != nil {
 		t.Error(err)
 	}
 
-	expectedOutput := `{"changes":[{"operation":"TUPLE_OPERATION_WRITE","timestamp":"2009-11-10T23:00:00Z","tuple_key":{"object":"document:doc1","relation":"reader","user":"user:user1"}}],"continuation_token":"01GXSA8YR785C4FYS3C0RTG7B2"}` //nolint:lll
+	expectedOutput := `{"changes":[{"operation":"TUPLE_OPERATION_WRITE","timestamp":"2009-11-10T23:00:00Z","tuple_key":{"object":"document:doc1","relation":"reader","user":"user:user1"}}],"continuation_token":"01GXSA8YR785C4FYS3C0RTG7B2"}`
 
 	outputTxt, err := json.Marshal(output)
 	if err != nil {